platform/requirements.txt
pulumi-azuread>=4.0.0,<5.0.0
pulumi-kubernetes>=3.0.0,<4.0.0